2.1.9修复网易地址 修复网易报错问题 删除Aria2下载  没用到这个功能后续需要再加上

+++ src/main/java/com/sqmusicplus/plug/netease/hander/NeteaseHander.java
@@ -108,7 +108,7 @@ public class NeteaseHander extends SearchHanderAbstract {
         plugSearchResult.setSearchIndex(searchKeyData.getPageIndex())
                 .setSearchSize(searchKeyData.getPageSize())
                 .setSearchType(searchKeyData.getSearchType())
-                .setSearchTotal( searchMusicResult.getResult().getSongCount())
+                .setSearchTotal( searchMusicResult.getResult().getSongCount().intValue())
                 .setSearchKeyWork(searchKeyData.getSearchkey())
                 .setRecords(plugSearchMusicResults);
         plugSearchResult.setSearchType(searchKeyData.getSearchType());
